+++ src/Recyclarr.Cli/Console/Commands/SyncCommand.cs
@@ -8,6 +8,7 @@ using Recyclarr.Cli.Migration;
 using Recyclarr.Cli.Processors;
 using Recyclarr.TrashLib.Config;
 using Recyclarr.TrashLib.Repo;
+using Recyclarr.TrashLib.Settings;
 using Spectre.Console.Cli;
 
 namespace Recyclarr.Cli.Console.Commands;
@@ -19,6 +20,7 @@ public class SyncCommand : AsyncCommand<SyncCommand.CliSettings>
     private readonly IMigrationExecutor _migration;
     private readonly IRepoUpdater _repoUpdater;
     private readonly ISyncProcessor _syncProcessor;
+    private readonly ISettingsProvider _settings;
 
     [UsedImplicitly]
     [SuppressMessage("Design", "CA1034:Nested types should not be visible")]
@@ -53,11 +55,13 @@ public class SyncCommand : AsyncCommand<SyncCommand.CliSettings>
     public SyncCommand(
         IMigrationExecutor migration,
         IRepoUpdater repoUpdater,
-        ISyncProcessor syncProcessor)
+        ISyncProcessor syncProcessor,
+        ISettingsProvider settings)
     {
         _migration = migration;
         _repoUpdater = repoUpdater;
         _syncProcessor = syncProcessor;
+        _settings = settings;
     }
 
     [SuppressMessage("Design", "CA1031:Do not catch general exception types")]
@@ -65,7 +69,7 @@ public class SyncCommand : AsyncCommand<SyncCommand.CliSettings>
     {
         // Will throw if migration is required, otherwise just a warning is issued.
         _migration.CheckNeededMigrations();
-        await _repoUpdater.UpdateRepo();
+        await _repoUpdater.UpdateRepo(_settings.Settings.Repositories.TrashGuide);
 
         return (int) await _syncProcessor.ProcessConfigs(settings);
     }