+++ src/Recyclarr.TrashLib/Config/Listers/ConfigTemplateLister.cs
@@ -1,5 +1,6 @@
 using Recyclarr.TrashLib.Config.Services;
 using Recyclarr.TrashLib.Repo;
+using Recyclarr.TrashLib.Settings;
 using Spectre.Console;
 
 namespace Recyclarr.TrashLib.Config.Listers;
@@ -9,20 +10,23 @@ public class ConfigTemplateLister : IConfigLister
     private readonly IAnsiConsole _console;
     private readonly IConfigTemplateGuideService _guideService;
     private readonly IRepoUpdater _repoUpdater;
+    private readonly ISettingsProvider _settings;
 
     public ConfigTemplateLister(
         IAnsiConsole console,
         IConfigTemplateGuideService guideService,
-        IRepoUpdater repoUpdater)
+        IRepoUpdater repoUpdater,
+        ISettingsProvider settings)
     {
         _console = console;
         _guideService = guideService;
         _repoUpdater = repoUpdater;
+        _settings = settings;
     }
 
     public async Task List()
     {
-        await _repoUpdater.UpdateRepo();
+        await _repoUpdater.UpdateRepo(_settings.Settings.Repositories.TrashGuide);
 
         var data = _guideService.TemplateData;