+++ src/Recyclarr.TrashLib/Repo/IRepoUpdater.cs
@@ -1,9 +1,10 @@
 using System.IO.Abstractions;
+using Recyclarr.TrashLib.Settings;
 
 namespace Recyclarr.TrashLib.Repo;
 
 public interface IRepoUpdater
 {
     IDirectoryInfo RepoPath { get; }
-    Task UpdateRepo();
+    Task UpdateRepo(IRepositorySettings repoSettings);
 }