+++ src/Recyclarr.TrashLib/Settings/IRepositorySettings.cs
@@ -1,0 +1,8 @@
+namespace Recyclarr.TrashLib.Settings;
+
+public interface IRepositorySettings
+{
+    Uri CloneUrl { get; }
+    string Branch { get; }
+    string? Sha1 { get; }
+}